requirements.txt
APScheduler==3.5.0
emoji==0.5.4
emoji-country-flag==1.2.1
Flask==1.0.2
Flask-APScheduler==1.11.0
gender-guesser==0.4.0
gevent==1.4.0
googletrans==2.4.0
python-telegram-bot==6.1.0
pandas==0.24.1
regex==2018.1.10
Telethon==1.11.3
gunicorn==20.0.4
schedule==0.6.0
requests==2.7.0



